Worked on data mocking, added functionality for calendar data and recipe data #28
